pyproject.toml
[tool.poetry]
name = "cohort-builder-data-dictionary"
version = "0.1.0"
description = "Data model for cohort builder"
authors = ["Nicholas Lee <nicholas.lee@sagebase.org>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9.0,<3.11"
pyyaml = "^6.0.1"
pandas = ">=1.3.1,<2.0.0"
numpy = "^1.25.2"
schematicpy = "^23.9.1"
types-pyyaml = "^6.0.12.11"
thefuzz = "^0.20.0"
requests = "^2.31.0"
toolbox = {git = "git@github.com:nlee-sage/toolbox.git"}


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.2"
mypy = "^1.5.1"
pylint = "^2.17.6"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
